Clarify Redis Options and Update Defaults
There are 3 proposed changes:

* Clarify the settings required between *Redis*, *Redis Sentinel* and *Redis Cluster*
* Add examples explicitly for "Redis Sentinel"
* Change default for `enableCluster` to `false`.

The latter could be a breaking change, however the `bitnami/redis` chart used in the documentation does not run as a cluster, therefore this proposal aligns with the documentation (https://tyk.io/docs/tyk-oss/ce-helm-chart/).  However, the documentation also references a "Redis Cluster" address so there is an inconsistency.  The choice is to either change the documentation to reflect a *Redis Cluster* (e.g. `bitnami/redis-cluster`) or use the existing Bitnami Redis chart (`bitnami/redis`) but then update the defaults in this `values.yaml` file. This change proposes the latter.