Clarify Redis Options and Update Defaults (#218)
* Clarify Redis Options and Update Defaults

There are 3 proposed changes:

* Clarify the settings required between *Redis*, *Redis Sentinel* and *Redis Cluster*
* Add examples explicitly for "Redis Sentinel"
* Change default for `enableCluster` to `false`.

The latter could be a breaking change, however the `bitnami/redis` chart used in the documentation does not run as a cluster, therefore this proposal aligns with the documentation (https://tyk.io/docs/tyk-oss/ce-helm-chart/).  However, the documentation also references a "Redis Cluster" address so there is an inconsistency.  The choice is to either change the documentation to reflect a *Redis Cluster* (e.g. `bitnami/redis-cluster`) or use the existing Bitnami Redis chart (`bitnami/redis`) but then update the defaults in this `values.yaml` file. This change proposes the latter.

* Update tyk-headless/values.yaml

Co-authored-by: Burak Sekili <buraksekili@gmail.com>

* Update tyk-headless/values.yaml

Co-authored-by: Burak Sekili <buraksekili@gmail.com>

* Update tyk-headless/values.yaml

Co-authored-by: Burak Sekili <buraksekili@gmail.com>

* Updated to reflect comments.

Addressed comments with the following changes:

* Added back in example using tyk simple redis chart.
* Reverted default to `enableCluster: true` as per original
* Minor formatting to make language consistent e.g. `sentinel` instead of `Sentinel`

* Update tyk-headless/values.yaml

Co-authored-by: Burak Sekili <buraksekili@gmail.com>

* Removed Trailing Spaces

* Update tyk-headless/values.yaml

Co-authored-by: Komal Sukhani <komaldsukhani@gmail.com>

* Update tyk-headless/values.yaml

Co-authored-by: Komal Sukhani <komaldsukhani@gmail.com>

* Changelog Updated for Redis Clarification

Co-authored-by: Burak Sekili <buraksekili@gmail.com>
Co-authored-by: Komal Sukhani <komaldsukhani@gmail.com>